pyproject.toml
[tool.poetry]
name = "3_robotframework"
version = "0.1.0"
description = ""
authors = ["martinstajnko <stajnko.martin93@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
robotframework = "^4.1.2"
robotframework-requests = "^0.9.2"
robotframework-seleniumlibrary = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
